[STABLE] Done with Regimens on frontend?
